Added support for SalesForce
SalesForce integration uses standard DBMS table integration. Some
changes were made to reflect limitations in SalesForce driver.

+++ test/src/org/tms/tds/dbms/DbmsTableImplTest.java
@@ -40,7 +40,7 @@ public class DbmsTableImplTest extends BaseDbmsTest
         ResultSet rs = fetchResultSet("jdbc:mysql://localhost/cds?user=davids&password=mysql", 
                                       "select * from employee order by name");
         
-        int numRows = getDbmsRowCount(rs);
+        int numRows = (int)getDbmsRowCount(rs);
         int numCols = getDbmsColumnCount(rs);
         close(rs);
         
@@ -119,7 +119,7 @@ public class DbmsTableImplTest extends BaseDbmsTest
         ResultSet rs = fetchResultSet("jdbc:mysql://localhost/music?user=davids&password=mysql", 
                                       "select * from metadata");
         
-        int numRows = getDbmsRowCount(rs);
+        int numRows = (int)getDbmsRowCount(rs);
         int numCols = getDbmsColumnCount(rs);
         close(rs);
